+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "files.associations": {
+        "iostream": "cpp",
+        "set": "cpp",
+        "map": "cpp"
+    }
+}
