+++ srcs/client/client.hpp
@@ -1,0 +1,2 @@
+#include "../../Inc/define.hpp"
+
